Add db storage (#12)
* add DB storage

* 1.Иправил тип chatId на Long 2. исправил UUID 3.Доделал dao db

* delet iml

* set 4 for poolSize

* set 2 pool

* 1. Fixed DbStorage 2.Fixed UUID 3.

* Fixed poolSize

* 1.Add transaction in method removeUser 2.Fixed loging

* Fixed weekDistance

* Delete logging.log

* Update Readme

* R

* Update README.md

* Update README.md

* Fixed TODO

* Update README.md

+++ src/main/java/ru/artak/service/TelegramUserInfo.java
@@ -2,17 +2,17 @@ package ru.artak.service;
 
 public class TelegramUserInfo {
 
-    private final Integer chatId;
+    private final Long chatId;
     private final String text;
     private final Integer updateId;
 
-    public TelegramUserInfo(Integer chatId, String text, Integer updateId) {
+    public TelegramUserInfo(Long chatId, String text, Integer updateId) {
         this.chatId = chatId;
         this.text = text;
         this.updateId = updateId;
     }
 
-    public Integer getChatId() {
+    public Long getChatId() {
         return chatId;
     }
 
@@ -20,5 +20,7 @@ public class TelegramUserInfo {
         return text;
     }
 
-    public Integer getUpdateId() { return updateId; }
+    public Integer getUpdateId() {
+        return updateId;
+    }
 }
